Provide support for Python 2.5.
Mac OSX Leopard is stuck on Python 2.5.1. This Python version does not support
the -B flag which is used to suppress the Python bytecode files and thus
refuses to run.

By adding a wrapper script which sets the environment variable, it is possible
to have the desired behavior on Python 2.6 and up, while still allowing Python
2.5 to be used.